various small refactorings
