+++ src/main/java/com/s24/search/solr/component/termstrategy/AdditiveTermRankingStrategy.java
@@ -1,6 +1,5 @@
 package com.s24.search.solr.component.termstrategy;
 
-import com.s24.search.solr.component.AbstractTermRankingStrategy;
 import org.apache.solr.common.params.ModifiableSolrParams;
 
 public class AdditiveTermRankingStrategy extends AbstractTermRankingStrategy {
